test: move tests related to Document methods to right file
